P_agentic/requirements.txt
langchain>=0.1.10
langgraph>=0.0.30
groq>=0.5.0
tavily-python>=0.3.0
streamlit>=1.30.0
pydantic>=2.5.0
python-dotenv>=1.0.0
reportlab>=4.0.0
fpdf2